requirements.txt
streamlit
pandas 
numpy 
yfinance 
plotly.express
scipy.optimize import minimize
datetime import datetime, timedelta